simple-tile: refactor per-wset data into separate file
